# Conflicts:
#	tools/StaticAnalysis/Exceptions/BreakingChangeIssues.csv

+++ src/ResourceManager/Network/Commands.Network/Models/PSVirtualNetworkUsage.cs
@@ -3,8 +3,8 @@
     public partial class PSVirtualNetworkUsage
     {
         public string Id { get; set; }
-        public long CurrentValue { get; set; }
-        public long Limit { get; set; }
+        public double CurrentValue { get; set; }
+        public double Limit { get; set; }
         public string Unit { get; set; }
         public PSUsageName Name { get; set; }
     }
